pyproject.toml
[project]
name = "bq-watchdog"
version = "0.1.0"
description = "AI-powered BigQuery cost guard for dbt projects"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["bigquery", "dbt", "cost", "data-engineering", "ci", "github-actions"]

dependencies = [
    "google-cloud-bigquery>=3.0",
    "sqlglot>=25.0",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "click>=8.0",
    "rich>=13.0",
    "PyGithub>=2.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "dbt-core", "dbt-bigquery"]

[project.scripts]
bqwatch  = "bq_watchdog.cli:cli"

[project.urls]
Repository = "https://github.com/yourusername/bq-watchdog"
"GitHub Marketplace" = "https://github.com/marketplace/actions/bq-watchdog"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bq_watchdog*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: requires live GCP credentials (skipped by default)",
]